tests: fix integration test file handling
The integration tests were failing with messages about build.gradle already existing.
This appears to be caused by the tests making the assumption that TemporaryFolder.newFolder/newFile
would return the same file if called repeatedly with the same arguments.  However, it looks like that was
changed such that now it throws an exception if the file exists.

The new implementation should meet the expectations of the existing tests.

https://github.com/junit-team/junit/issues/413
https://github.com/junit-team/junit/pull/419