Pass grailsHome to launcher via system property when available
Grails depends on grails.home being set for some features.  This plugin already
had some support for providing a grailsHome via project property (which can
either be set on the command-line or by assigning a value to "ext.grailsHome" in
the build script), but this value wasn't being passed to the launcher so that it can
be used.  In particular, after this change, if you set the grailsHome project property,
when running the grails-run-app task, Environment.isDevelopmentMode() should
now return true.